+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - com.dwarfeng.subgrade.impl.service.AbstractCrudService。
   - com.dwarfeng.subgrade.impl.service.AbstractEntireLookupService。
   - com.dwarfeng.subgrade.impl.service.AbstractPresetLookupService。
+  - com.dwarfeng.subgrade.impl.service.AbstractBatchWriteService。
+  - com.dwarfeng.subgrade.impl.service.AbstractWriteService。
 
 - 优化部分服务的实现的继承关系。
   - com.dwarfeng.subgrade.impl.service.CustomBatchCrudService。
@@ -20,6 +22,8 @@
   - com.dwarfeng.subgrade.impl.service.GeneralEntireLookupService。
   - com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService。
   - com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService。
+  - com.dwarfeng.subgrade.impl.service.DaoOnlyBatchWriteService。
+  - com.dwarfeng.subgrade.impl.service.DaoOnlyWriteService。
 
 - 优化部分接口及其实现的参数名称。
   - com.dwarfeng.subgrade.stack.service.BatchCrudService。